fix: 旧スキーマの.cliptapを移行してから取り込むようにする
インポートおよびWeb初期画面の初回ファイル読込で、取り込むDBのスキーマ版が
最新(V7)と異なる場合に移行してから反映し、版の相違でエラーにならないようにする。
併せて、目的に対して過剰だった実装と冗長コードを整理する。

## 移行

- Web初回読込を ImportService.prepareImportDatabase 経由に統一し、検証と移行を
  一時DB上で完了してからメインDBへ昇格する。移行途中のDBでワークスペースを開かない
- 移行段の選択は宣言バージョンだけで行い、移行後は必須テーブルの存在だけを確認する。
  テーブル名は CREATE_TABLES のDDLから導出し、スキーマ定義との二重管理を作らない
- 旧実装で作りそこねた派生インデックスを移行完了時に再作成する
- V3〜V7がインポート・起動時移行の双方でV7へ移行できることを回帰テストで固定する

## リリース済みマイグレーションの差し戻し

V4→V5・V5→V6に加えられていた「冪等化」は sortOrder を無条件に全行再計算するため、
systemDBの版が実DBより過小な状態で走ると利用者の並び順を破壊する。公開済みの実装へ
戻し、「列が既にある場合は並び順に触れない」回帰テストを追加する。

## 過剰実装の削減

- Web Locksによるタブ排他、SingleFlight、valid更新の保留機構、
  snapshot/rollback機構、キャッシュ世代管理、closeMobileDatabase.ts を削除
- 初回読込の失敗時は database.reset() で未読込状態へ戻す。到達時点で
  キャッシュへ書き込まないため、読込前の状態がそのまま残る
- 差し替え中は自動キャッシュ保存を停止し、システムDBを欠いたキャッシュを残さない

## 不具合修正

- MobileDatabaseAdapter.close() が closeSync() の例外で参照を残さないようにする
- 初回読込の巻き戻し失敗が本来のエラーを置き換えないようにする
- 起動時の移行中も自動保存を停止し、中断したDBがキャッシュへ焼き付かないようにする
- 内部Errorの英語メッセージを画面へ出さず、翻訳キーで表示する

## 文書

- 実在しない「第15章 未確定事項」への参照を全廃し、機能仕様書§1.1の運用へ揃える
- README/DEVELOPMENT の古い版数、存在しないと書かれていた npm test を修正
- 旧版starterファイルを配信対象から外し、移行回帰テストのfixtureとして保持する

+++ packages/shared/src/types/export.ts
@@ -165,7 +165,7 @@ export type ImportCandidates = z.infer<typeof ImportCandidatesSchema>;
  * - c: チェックサム（SHA-256、改竄検知）
  */
 export const ClipTapExportDataSchema = z.object({
-  s: z.number(),
+  s: z.number().int(),
   t: z.string(),
   h: z.string(),
   d: z.string(),
